started re-writing visual using D3.js.
I decided to switch to a different library for rendering the visual. It became clear that the previous one (cytoscape.js) did not offer enough flexibility to implement all the features that we want. D3.js is the most feature rich data visualization library, so I think it will work well for what we are trying to do.